requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.6
django-import-export==4.3.9
djangorestframework==3.16.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
django-static-fontawesome
django-static-jquery3
django-admin-global-sidebar
django-rangefilter
django-slick-reporting
django-crispy-forms
crispy-bootstrap5
xhtml2pdf
